Add function heterozygosity_blue following Harris and DeGiorgio (2017)
This function calculates expected heterozygosity with a correction
for relatedness of individuals based on a kinship matrix.
In this implementation the kinship matrix is an optional argument
an if it is not present then the 'typical' expected heterozygosity
is calculated which is equivilent to heterozygosity_blue with
unrelated individuals.
This function is also suitable for mixed ploidy data.